Przeniesiono kod obsługi bazy danych do oddzielnej klasy.
